+++ Makefile
@@ -2,7 +2,7 @@ NAME = minishell
 SRCDIR = srcs
 OBJSDIR = objs
 SRCS = \
-	srcs/builtins/is_builtin.c srcs/builtins/my_cd.c srcs/builtins/my_cd_utils.c srcs/builtins/my_echo.c srcs/builtins/my_env.c srcs/builtins/my_execve.c srcs/builtins/my_exit.c srcs/builtins/my_exit_help.c srcs/builtins/my_export.c srcs/builtins/my_export_error.c srcs/builtins/my_export_modify_env_value.c srcs/builtins/my_export_to_string_parse_token.c srcs/builtins/my_pwd.c srcs/builtins/my_unset.c srcs/env/expand_env.c srcs/env/token_lst.c srcs/error/check_syntax_error.c srcs/exec/exec_one_cmd.c srcs/exec/main.c srcs/exec/pipefds.c srcs/lexer/add_colon_doll_token.c srcs/lexer/add_pipe_redirect_token.c srcs/lexer/add_quote_token.c srcs/lexer/lexer.c srcs/parser/get_cmd_args.c srcs/parser/get_redirect.c srcs/parser/parser.c srcs/parser/parser_utils.c srcs/parser/set_stat_content.c srcs/signal/handle_signal.c srcs/utils/change_finish_stat.c srcs/utils/free_array.c srcs/utils/free_lists.c srcs/utils/print_arr.c srcs/utils/print_list.c srcs/utils/utils_print_list.c 
+	srcs/builtins/is_builtin.c srcs/builtins/my_cd.c srcs/builtins/my_cd_utils.c srcs/builtins/my_echo.c srcs/builtins/my_env.c srcs/builtins/my_execve.c srcs/builtins/my_exit.c srcs/builtins/my_exit_help.c srcs/builtins/my_export.c srcs/builtins/my_export_error.c srcs/builtins/my_export_modify_env_value.c srcs/builtins/my_export_to_string_parse_token.c srcs/builtins/my_pwd.c srcs/builtins/my_unset.c srcs/env/expand_env.c srcs/env/token_lst.c srcs/error/check_syntax_error.c srcs/exec/envlist_to_envp.c srcs/exec/exec_in_child.c srcs/exec/exec_one_cmd.c srcs/exec/get_path.c srcs/exec/main.c srcs/exec/pipe.c srcs/exec/pipefds.c srcs/exec/redirect.c srcs/exec/redirect_in.c srcs/exec/redirect_out.c srcs/lexer/add_colon_doll_token.c srcs/lexer/add_pipe_redirect_token.c srcs/lexer/add_quote_token.c srcs/lexer/lexer.c srcs/parser/get_cmd_args.c srcs/parser/get_redirect.c srcs/parser/parser.c srcs/parser/parser_utils.c srcs/parser/set_stat_content.c srcs/signal/handle_signal.c srcs/utils/change_finish_stat.c srcs/utils/free_array.c srcs/utils/free_lists.c srcs/utils/print_arr.c srcs/utils/print_list.c srcs/utils/utils_print_list.c
 
 OBJS := $(patsubst srcs/%.c, objs/%.o, $(SRCS))
 LDFLAGS = -Llibft -lft -L $(shell brew --prefix readline)/lib -lreadline
